Finally starting to get bond futures passing the tests.

+++ projects/OG-Analytics/src/com/opengamma/financial/analytics/securityconverters/StubCalculator.java
@@ -49,13 +49,13 @@ public final class StubCalculator {
    * @param isEndOfMonthConvention  whether to use end of month rules
    * @return the stub type, not null
    */
-  public static StubType getStartStubType(final DateProvider[] schedule, final int paymentsPerYear, final boolean isEndOfMonthConvention) {
+  public static StubType getStartStubType(final DateProvider[] schedule, final double paymentsPerYear, final boolean isEndOfMonthConvention) {
     Validate.notNull(schedule, "schedule");
     Validate.noNullElements(schedule, "schedule");
     Validate.isTrue(paymentsPerYear > 0);
     Validate.isTrue(12 % paymentsPerYear == 0);
-    
-    final int months = 12 / paymentsPerYear;
+
+    final int months = (int) (12 / paymentsPerYear);
     final LocalDate first = LocalDate.of(schedule[0]);
     final LocalDate second = LocalDate.of(schedule[1]);
     LocalDate date;
@@ -100,13 +100,13 @@ public final class StubCalculator {
    * @param isEndOfMonthConvention  whether to use end of month rules
    * @return the stub type, not null
    */
-  public static StubType getEndStubType(final DateProvider[] schedule, final int paymentsPerYear, final boolean isEndOfMonthConvention) {
+  public static StubType getEndStubType(final DateProvider[] schedule, final double paymentsPerYear, final boolean isEndOfMonthConvention) {
     Validate.notNull(schedule, "schedule");
     Validate.noNullElements(schedule, "schedule");
     Validate.isTrue(paymentsPerYear > 0);
     Validate.isTrue(12 % paymentsPerYear == 0);
-    
-    final int months = 12 / paymentsPerYear;
+
+    final int months = (int) (12 / paymentsPerYear);
     final int n = schedule.length;
     final LocalDate first = LocalDate.of(schedule[n - 2]);
     final LocalDate second = LocalDate.of(schedule[n - 1]);
