+++ projects/OG-Analytics/src/com/opengamma/financial/convention/daycount/ThirtyThreeSixtyTypeDayCount.java
@@ -13,7 +13,7 @@ import javax.time.calendar.ZonedDateTime;
 public abstract class ThirtyThreeSixtyTypeDayCount extends StatelessDayCount {
 
   @Override
-  public double getAccruedInterest(final ZonedDateTime previousCouponDate, final ZonedDateTime date, final ZonedDateTime nextCouponDate, final double coupon, final int paymentsPerYear) {
+  public double getAccruedInterest(final ZonedDateTime previousCouponDate, final ZonedDateTime date, final ZonedDateTime nextCouponDate, final double coupon, final double paymentsPerYear) {
     return coupon * getDayCountFraction(previousCouponDate, date);
   }
