+++ projects/OG-Financial/tests/unit/com/opengamma/financial/analytics/schedule/ScheduleCalculatorTest.java
@@ -197,7 +197,7 @@ public class ScheduleCalculatorTest {
       }
 
       @Override
-      public double getAccruedInterest(final ZonedDateTime previousCouponDate, final ZonedDateTime date, final ZonedDateTime nextCouponDate, final double coupon, final int paymentsPerYear) {
+      public double getAccruedInterest(final ZonedDateTime previousCouponDate, final ZonedDateTime date, final ZonedDateTime nextCouponDate, final double coupon, final double paymentsPerYear) {
         return 0;
       }
